requirements.txt
streamlit==1.32.0
ultralytics==8.1.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
matplotlib==3.8.3
tqdm==4.66.2
seaborn==0.13.2
pandas==2.2.1